change the delete logic, remove console log
